requirements.txt
fastapi
numpy
scikit-learn
uvicorn
gunicorn
gTTS
tensorflow
pathlib
Pillow
starlette
python-multipart